update fuel data

+++ src/main/java/com/example/cellfire/data/ForestTypeConditions.java
@@ -1,7 +1,7 @@
 package com.example.cellfire.data;
 
 public final class ForestTypeConditions {
-    public static double determineIgnitionTemperature(byte forestType) {
+    public static double determineIgnitionTemperature(int forestType) {
         return switch (forestType) {
             case 1 -> 325;
             case 2 -> 275;
@@ -12,7 +12,7 @@ public final class ForestTypeConditions {
         };
     }
 
-    public static double determineActivationEnergy(byte forestType) {
+    public static double determineActivationEnergy(int forestType) {
         return switch (forestType) {
             case 1 -> 125_000;
             case 2 -> 105_000;
@@ -24,11 +24,11 @@ public final class ForestTypeConditions {
     }
 
     public static final class ForestType {
-        public static final byte TREELESS  = 0;
-        public static final byte EVERGREEN_NEEDLE_LEAF = 1;
-        public static final byte EVERGREEN_BROADLEAF = 2;
-        public static final byte DECIDUOUS_NEEDLE_LEAF = 3;
-        public static final byte DECIDUOUS_BROADLEAF = 4;
-        public static final byte MIXED = 5;
+        public static final int TREELESS  = 0;
+        public static final int EVERGREEN_NEEDLE_LEAF = 1;
+        public static final int EVERGREEN_BROADLEAF = 2;
+        public static final int DECIDUOUS_NEEDLE_LEAF = 3;
+        public static final int DECIDUOUS_BROADLEAF = 4;
+        public static final int MIXED = 5;
     }
 }
